requirements.txt
openai>=1.37.1
python-dotenv>=1.0.1
